1.添加代码处理无法解析的pdf

+++ pdf2html.py
@@ -58,14 +58,38 @@ while True:
 		# 	c.kill()
 		for proc in psutil.process_iter():
 			if proc.name() == 'Acrobat.exe':
-				if(proc.memory_info().rss / 1024 /1024 > 2000):
-				# 测试使用120
-				# if(proc.memory_info().rss / 1024 /1024 > 120):
-				# if(proc.memory_info().rss > 104857600):
-				# if(proc.memory_info().rss):
-					print(proc.memory_info().rss)
-					sub_c.kill()
-					p=os.system("taskkill /f /im Acrobat.exe")
+				mem_info=int(proc.memory_info().rss / 1024 /1024)
+				time.sleep(1)
+				mem_tmp=int(proc.memory_info().rss / 1024 /1024)
+				a=1
+				while True:
+					mem_info=int(proc.memory_info().rss / 1024 /1024)
+					time.sleep(1)
+					print(a)
+					print('目前内存占用: %dM' % mem_info)
+					mem_tmp=int(proc.memory_info().rss / 1024 /1024)
+					if not abs(mem_tmp-mem_info):
+						a+=1
+					else:
+						a=0
+					if a > 60:
+						print('目前内存占用: %dM' % mem_info)
+						sub_c.kill()
+						p=os.system("taskkill /f /im Acrobat.exe")
+						print('该pdf无法解析，Acrobat进程已kill')
+						break
+					if(mem_info > 2000):
+					# 测试使用120
+					# if(proc.memory_info().rss / 1024 /1024 > 120):
+					# if(proc.memory_info().rss > 104857600):
+					# if(proc.memory_info().rss):
+						print('目前内存占用: %dM' % mem_info)
+						sub_c.kill()
+						p=os.system("taskkill /f /im Acrobat.exe")
+						print('该pdf解析使用内存过多，Acrobat进程已kill')
+						break
+		if a > 60:
+			break
 		time.sleep(2)
 		if p == 0:
 			break
@@ -77,20 +101,23 @@ while True:
 	print('解析花费: %s 分钟' % spend_t)
 	html_file_src=glob.glob("%s/*.html" % output)
 	html_length=len(html_file_src)
-	if p !=0:
-		html_list=html_file_src[0].split("\\")[1]
-		if html_length > 0:
-			sub_c.kill()
-			shutil.move("%s/%s" % (output,html_list),"%s/%s" % (folder,html_list))
-		
-		#移动文件计1s
-		time.sleep(2)
-		# 如果有html就不进入下一个循环
-		html_file=glob.glob("%s/*.html" % folder)
-		html_length=len(html_file)
-		while html_length:
+	if a < 60:
+		if p !=0:
+			html_list=html_file_src[0].split("\\")[1]
+			if html_length > 0:
+				sub_c.kill()
+				shutil.move("%s/%s" % (output,html_list),"%s/%s" % (folder,html_list))
+			
+			#移动文件计1s
+			time.sleep(2)
+			# 如果有html就不进入下一个循环
 			html_file=glob.glob("%s/*.html" % folder)
 			html_length=len(html_file)
+			while html_length:
+				html_file=glob.glob("%s/*.html" % folder)
+				html_length=len(html_file)
+		else:
+			break
 	else:
 		break
